requirements.txt
# CLI & Configuration
click>=8.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Audio Processing
pydub>=0.25.0
ffmpeg-python>=0.2.0

# Audio Recording
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0

# HTTP & APIs
httpx>=0.24.0
tenacity>=8.2.0
openai>=1.0.0

# Output & Formatting
rich>=13.0.0
jinja2>=3.1.0

# Utilities
python-dateutil>=2.8.0
loguru>=0.7.0
typing-extensions>=4.5.0

# Development Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.0.280
mypy>=1.4.0
